+ refactor form previous coursework
